requirements.txt
fastapi
uvicorn
pandas
scikit-learn
streamlit
kafka-python
dbt-core
dbt-postgres
pytest
kafka-python
psycopg2-binary
fastapi
uvicorn[standard]
requests